# testing: reduce memory allocation in Helper