@bboreham 2020-05-03
Stems from me discovering that calling `testify/require.NoError()` in the middle of a tight loop caused more memory allocation than the code it was testing.

Some discussion and a benchmark at https://github.com/stretchr/testify/pull/933#issuecomment-620217067